requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.8.1
packaging==24.2
regex==2024.11.6
setuptools==78.1.0
tqdm==4.67.1
werkzeug==2.3.6
cryptography==42.0.5
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0  # Add this line
openai==1.73.0